Update/profile (#9)
* Removed duplicate app paths

* Corrected container width for heroes and added separate view for new babble.

* Reviewed README

* Profile page remake in progress.

* Added babbles that have the tag when searching for keyword in the searchbar

* Making loging easier (works when CRUSHING enter on form)

* Changed folder structure, spacing in characters of babbles and new ER Diag based on this branch

* Fixed paths in populate to make it work again

* Fixed readme, changed password for DB

* Fixed start index on populate.pyX

* profile and myprofile complete and fully functional

* Changed password.. kind of over complicated

* Added is-square attribute for profile cards for better unicity.

+++ src/app.py
@@ -5,8 +5,8 @@ import os
 import shutil
 
 from flask import Flask, render_template, request, session, redirect
-from src.data.babblerdb import BabblerDB
-from src.data.utils import crop_tags_in_message
+from src.babbler.babblerdb import BabblerDB
+from src.babbler.utils import crop_tags_in_message
 
 
 app = Flask(__name__)
@@ -18,7 +18,7 @@ app.secret_key = 'YmxhemVpdDQyMA=='
 # Config data for DB
 app.config['DB_HOST'] = 'localhost'
 app.config['DB_USER'] = 'root'
-app.config['DB_PASSWORD'] = 'choupi'
+app.config['DB_PASSWORD'] = 'babblerisawesome'
 app.config['DB_NAME'] = 'Babbler'
 db = BabblerDB(app)
 
@@ -56,14 +56,17 @@ def new_babble():
         return redirect('/login')
 
 
-@app.route('/delete/<user>', methods=['POST'])
+@app.route('/delete/<user>', methods=['GET', 'POST'])
 def delete(user):
     connected_user = 'username' in session
-    if connected_user and session['username'] == user:
-        session.pop('username', None)
-        db.remove_babbler(user)
-        os.remove('./static/images/' + str(user) + '.jpg')
-        return redirect('/login')
+    if request.method == 'POST':
+        if connected_user and session['username'] == user:
+            session.pop('username', None)
+            db.remove_babbler(user)
+            os.remove('./static/images/' + str(user) + '.jpg')
+            return redirect('/login')
+    else:
+        return render_template('/partials/delete_user.html', user=user, logged=True)
 
 
 @app.route('/follow/<other>', methods=['POST'])
@@ -123,7 +126,18 @@ def my_profile():
     if 'username' in session:
         user = {'username': session['username'], 'publicName': session['publicName']}
         babbles = db.read_user_babbles(user['username'])
-        return render_template('/partials/myprofile.html', user=user, babbles=babbles, logged=True)
+        followers = db.read_followers(user['username'])
+        subscriptions = db.read_subscriptions(user['username'])
+        followed = db.following(session['username'], user['username'])
+        nb_babbles = len(babbles)
+        nb_followers = len(followers)
+        nb_subscriptions = len(subscriptions)
+        view = render_template('/partials/myprofile.html', user=user,
+                               followed=followed, babbles=babbles, logged=True,
+                               followers=followers, subscriptions=subscriptions,
+                               nb_babbles=nb_babbles, nb_followers=nb_followers,
+                               nb_subscriptions=nb_subscriptions)
+        return view
     else:
         return redirect('/login')
 
@@ -143,9 +157,18 @@ def babbler_profile(username):
         if session['username'] == username:
             return redirect('/myprofile')
         user = db.read_babblers(username)
-        followed = db.following(session['username'], username)
         babbles = db.read_user_babbles(username)
-        view = render_template('/partials/profile.html', user=user, followed=followed, babbles=babbles, logged=True)
+        followers = db.read_followers(username)
+        subscriptions = db.read_subscriptions(username)
+        followed = db.following(session['username'], username)
+        nb_babbles = len(babbles)
+        nb_followers = len(followers)
+        nb_subscriptions = len(subscriptions)
+        view = render_template('/partials/profile.html', user=user[0],
+                               followed=followed, babbles=babbles, logged=True,
+                               followers=followers, subscriptions=subscriptions,
+                               nb_babbles=nb_babbles, nb_followers=nb_followers,
+                               nb_subscriptions=nb_subscriptions)
         return view
     else:
         return redirect('/login')
@@ -168,6 +191,8 @@ def search_form():
         if keyword:
             babbles = db.read_babbles(keyword)
             babblers = db.read_babblers(keyword)
+            babbles_with_tag = db.read_babbles_with_tag(keyword)
+            babbles.extend(babbles_with_tag)
             return render_template('/partials/search_results.html',
                                    keyword=keyword, babbles=babbles,
                                    babblers=babblers, logged=True)
